// app/[locale]/(routes)/crm/opportunities/table-data/schema.tsx
import { z } from "zod";

// We're keeping a simple non-relational schema here.
// IRL, you will have a schema for your data models.
export const opportunitySchema = z.object({
  //TODO: fix all the types and nullable
  id: z.string(),
  name: z.string(),
  description: z.string(),
  next_step: z.string().nullable(),
  close_date: z.date().nullable().optional(),
  status: z.string().nullable(),
  budget: z.union([z.number(), z.bigint()]).nullable().transform((val) =>
    typeof val === 'bigint' ? Number(val) : val
  ),
  expected_revenue: z.union([z.number(), z.bigint()]).nullable().transform((val) =>
    typeof val === 'bigint' ? Number(val) : val
  ),
  assigned_account: z.object({}).optional().nullable(),
  assigned_sales_stage: z.object({}).optional().nullable(),
  assigned_to_user: z.object({}).optional().nullable(),
});

export type Opportunity = z.infer<typeof opportunitySchema>;